# Conflicts:
#	app/Http/Traits/ProductTrait.php

+++ database/migrations/2023_07_26_064222_alter_table_products.php
@@ -14,9 +14,15 @@ class AlterTableProducts extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->decimal('height', 10, 4)->nullable();
-            $table->decimal('breadth', 10, 4)->nullable();
-            $table->decimal('length', 10, 4)->nullable();
+            if(!Schema::hasColumn('products', 'height')){
+                $table->decimal('height', 10, 4)->nullable();
+            }
+            if(!Schema::hasColumn('products', 'breadth')){
+                $table->decimal('breadth', 10, 4)->nullable();
+            }
+            if(!Schema::hasColumn('products', 'length')){
+                $table->decimal('length', 10, 4)->nullable();
+            }
         });
     }
 
@@ -28,9 +34,15 @@ class AlterTableProducts extends Migration
     public function down()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('height');
-            $table->dropColumn('breadth');
-            $table->dropColumn('length');
+            if(Schema::hasColumn('products', 'height')){
+                $table->dropColumn('height');
+            }
+            if(Schema::hasColumn('products', 'breadth')){
+                $table->dropColumn('breadth');
+            }
+            if(Schema::hasColumn('products', 'length')){
+                $table->dropColumn('length');
+            }
         });
     }
 }
